Split the getWalletAccount
