+++ Project/scripts/deployments.ts
@@ -4,10 +4,6 @@ import "dotenv/config";
 import * as ballotJson from "../artifacts/contracts/CustomBallot.sol/CustomBallot.json";
 import * as tokenJson from "../artifacts/contracts/Token.sol/MyToken.json";
 
-// This key is already public on Herong's Tutorial Examples - v1.03, by Dr. Herong Yang
-// Do never expose your keys like this
-const EXPOSED_KEY =
-  "8da4ef21b864d2cc526dbdb2a120bd2874c36c9d0a1fb7f8c63d7f7a8b41de8f";
 
 function convertStringArrayToBytes32(array: string[]) {
   const bytes32Array = [];
@@ -17,7 +13,7 @@ function convertStringArrayToBytes32(array: string[]) {
   return bytes32Array;
 }
 
-async function main(...proposals: string[]) {
+async function main(signer: ethers.Wallet, ...proposals: string[]) {
   // =================================================================
   // Check the proposals
   // =================================================================
@@ -26,16 +22,6 @@ async function main(...proposals: string[]) {
   proposals.forEach((element, index) => {
     console.log(`Proposal N. ${index + 1}: ${element}`);
   });
-  // =================================================================
-  // Get Wallet Account
-  // =================================================================
-  const wallet =
-    process.env.MNEMONIC && process.env.MNEMONIC.length > 0
-      ? ethers.Wallet.fromMnemonic(process.env.MNEMONIC)
-      : new ethers.Wallet(process.env.PRIVATE_KEY ?? EXPOSED_KEY);
-  console.log(`Using address ${wallet.address}`);
-  const provider = ethers.providers.getDefaultProvider("rinkeby");
-  const signer = wallet.connect(provider);
   // =================================================================
   // Check Wallet Balance
   // =================================================================
